Remove OrcaRouter provider class per review feedback
The OrcaRouter gateway exposes an OpenAI-compatible API, and the separate
OrcaRouter class (plus the OpenAICompatibleClient base-class refactor it
required) adds no functionality over declaring additional components of
the existing OpenAI client with a distinct id and endpoint/model config.
As suggested in review, remove the new class and instead document how to
use the OpenAI client for OpenAI-compatible providers such as OrcaRouter
in the OpenAI class javadoc.
